Changed world object system
World objects are now based on layers. This means they only need to be drawn once and can then be moved at will without a significant performance impact (as opposed to redrawing everything slightly more to the right).

The system that joins the different layers together should probably be more optimized (especially in situations where one object clearly overwrites pixels below it - if a world object is completely opaque, there is no reason to draw the things below it since they will be covered anyway)

Changed most functions and values from u32 to usize to prevent having to 'as usize' all the values before indexing a vec.

+++ src/lib.rs
@@ -60,7 +60,7 @@ pub fn main_js() -> Result<(), JsValue> {
         let context: RenderingContext = context.dyn_into().unwrap();
         // context.clear_color(0.0, 0.0, 0.0, 1.0);
         // context.clear(RenderingContext::COLOR_BUFFER_BIT);
-        renderer::init_renderer(context, pixel_width_i, pixel_height_i, window, document);
+        renderer::init_renderer(context, pixel_width_i as usize, pixel_height_i as usize, window, document);
     }
 
     Ok(())
